move dataclasses to a dedicated place
